move blocklist to its own const file
