+++ src/nft/utils/index.ts
@@ -3,6 +3,7 @@ export * from './buildActivityAsset'
 export * from './buildSellObject'
 export * from './carousel'
 export * from './collection'
+export * from './constants'
 export * from './currency'
 export * from './formatEventProperties'
 export * from './isAudio'
